refactor(payments/shop): move rename offer lib folder to shop
