+++ packages/backend-modules/payments/lib/shop/index.ts
@@ -1,0 +1,3 @@
+export * from './offers'
+export * from './Shop'
+export * as utils from './utils'
